Update README.md
To run the Node server, navigate to the server directory and execute "npm run dev". This launches the backend environment, leveraging Express.js and MongoDB for data management. Concurrently, start the frontend server by moving to the client/reactcrud directory and running "npm run dev". This initiates the React-based frontend, facilitating seamless interaction with the CRUD API. For detailed instructions and code, visit the GitHub repository.